Move customProperties logic to the Customer model
