+++ sdk/src/main/java/com/talkable/sdk/models/Customer.java
@@ -6,12 +6,14 @@ import com.google.gson.annotations.SerializedName;
 import com.talkable.sdk.interfaces.ApiSendable;
 
 import java.io.UnsupportedEncodingException;
+import java.util.HashMap;
 
 public class Customer implements ApiSendable {
     @SerializedName("customer_id") String id;
     @SerializedName("first_name") String firstName;
     @SerializedName("last_name") String lastName;
     String email, encodedEmail;
+    @SerializedName("custom_properties") HashMap<String, String> customProperties;
 
     public Customer() {
 
@@ -22,10 +24,16 @@ public class Customer implements ApiSendable {
     }
 
     public Customer(String id, String firstName, String lastName, String email) throws UnsupportedEncodingException {
+        this(null, null, null, email, null);
+    }
+
+    public Customer(String id, String firstName, String lastName, String email, HashMap<String, String> customProperties) throws UnsupportedEncodingException {
         this.firstName = firstName;
+
         this.lastName = lastName;
         this.id = id;
         setEmail(email);
+        setCustomProperties(customProperties);
     }
 
     public String getId() {
@@ -69,5 +77,13 @@ public class Customer implements ApiSendable {
 
     public void setEncodedEmail(String encodedEmail) {
         this.encodedEmail = encodedEmail;
+    }
+
+    public HashMap<String, String> getCustomProperties() {
+        return customProperties;
+    }
+
+    public void setCustomProperties(HashMap<String, String> customProperties) {
+        this.customProperties = customProperties;
     }
 }
