+++ sdk/src/main/java/com/talkable/sdk/utils/UriUtils.java
@@ -88,6 +88,15 @@ public class UriUtils {
             params.put("o[email]", encodedEmail);
             params.put("o[first_name]", customer.getFirstName());
             params.put("o[last_name]", customer.getLastName());
+
+            Map<String, String> customProperties = customer.getCustomProperties();
+            if (customProperties != null) {
+                for (Map.Entry<String, String> prop : customProperties.entrySet()) {
+                    if (prop.getValue() != null) {
+                        params.put(String.format("custom_properties[%s]", prop.getKey()), prop.getValue());
+                    }
+                }
+            }
         }
 
         if (origin instanceof Event) {
@@ -118,15 +127,6 @@ public class UriUtils {
             } else {
                 params.put("o[event_number]", event.getEventNumber());
                 params.put("o[event_category]", event.getEventCategory());
-            }
-        }
-
-        Map<String, String> customProperties = origin.getCustomProperties();
-        if (customProperties != null) {
-            for (Map.Entry<String, String> prop : customProperties.entrySet()) {
-                if (prop.getValue() != null) {
-                    params.put(String.format("custom_properties[%s]", prop.getKey()), prop.getValue());
-                }
             }
         }
